Finally got liquibase to work the way I want it to.
Can now successfully execute ./gradlew generateChangeLog to generate a change log file based on the java implementation via connecting to hibernate as a referenceUrl.

+++ src/main/java/com/reading/tvirdee/project/springbootdockermysql/repository/UserRepository.java
@@ -1,9 +1,9 @@
 package com.reading.tvirdee.project.springbootdockermysql.repository;
 
-import com.reading.tvirdee.project.springbootdockermysql.entity.Users;
+import com.reading.tvirdee.project.springbootdockermysql.domain.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<Users, Integer> {
+public interface UserRepository extends JpaRepository<Users, Long> {
 }
